Support 64-bit core ID and hence aarch64 MPIDR register

+++ arch/arm/system_registers.c
@@ -73,9 +73,9 @@ uint64_t HELPER(get_cp_reg64)(CPUState * env, void *rip)
     return res;
 }
 
-static inline uint32_t get_mpidr(CPUState *env)
+static inline uint64_t get_mpidr(CPUState *env)
 {
-    int mpidr = tlib_get_mp_index();
+    uint64_t mpidr = tlib_get_mp_index();
     /* We don't support setting cluster ID ([8..11])
      * so these bits always RAZ.
      */
